+++ sdk/go/google/cloudresourcemanager/v3/project.go
@@ -30,7 +30,7 @@ type Project struct {
 	// Optional. A reference to a parent Resource. eg., `organizations/123` or `folders/876`.
 	Parent pulumi.StringOutput `pulumi:"parent"`
 	// Immutable. The unique, user-assigned id of the project. It must be 6 to 30 lowercase ASCII letters, digits, or hyphens. It must start with a letter. Trailing hyphens are prohibited. Example: `tokyo-rain-123`
-	Project pulumi.StringOutput `pulumi:"project"`
+	ProjectId pulumi.StringOutput `pulumi:"projectId"`
 	// The project lifecycle state.
 	State pulumi.StringOutput `pulumi:"state"`
 	// The most recent time this resource was modified.
@@ -83,7 +83,7 @@ type projectArgs struct {
 	// Optional. A reference to a parent Resource. eg., `organizations/123` or `folders/876`.
 	Parent *string `pulumi:"parent"`
 	// Immutable. The unique, user-assigned id of the project. It must be 6 to 30 lowercase ASCII letters, digits, or hyphens. It must start with a letter. Trailing hyphens are prohibited. Example: `tokyo-rain-123`
-	Project *string `pulumi:"project"`
+	ProjectId *string `pulumi:"projectId"`
 }
 
 // The set of arguments for constructing a Project resource.
@@ -95,7 +95,7 @@ type ProjectArgs struct {
 	// Optional. A reference to a parent Resource. eg., `organizations/123` or `folders/876`.
 	Parent pulumi.StringPtrInput
 	// Immutable. The unique, user-assigned id of the project. It must be 6 to 30 lowercase ASCII letters, digits, or hyphens. It must start with a letter. Trailing hyphens are prohibited. Example: `tokyo-rain-123`
-	Project pulumi.StringPtrInput
+	ProjectId pulumi.StringPtrInput
 }
 
 func (ProjectArgs) ElementType() reflect.Type {